improved error handling for system test server startup
Capture stderr and stdout when process is dead; only kill if it isn't
already dead.

Patch by Sammy Yu; reviewed by eevans for CASSANDRA-306

git-svn-id: https://svn.apache.org/repos/asf/incubator/cassandra/trunk@794951 13f79535-47bb-0310-9956-ffa450edef68

+++ test/system/__init__.py
@@ -72,7 +72,7 @@ class CassandraTester(object):
             os.chdir(root)
             os.environ['CASSANDRA_INCLUDE'] = 'test/cassandra.in.sh'
             args = ['bin/cassandra', '-p', pid_fname]
-            sp.Popen(args, stderr=sp.PIPE, stdout=sp.PIPE)
+            process = sp.Popen(args, stderr=sp.PIPE, stdout=sp.PIPE)
             time.sleep(0.1)
 
             # connect to it, with a timeout in case something went wrong
@@ -85,8 +85,16 @@ class CassandraTester(object):
                 else:
                     break
             else:
-                os.kill(pid(), signal.SIGKILL) # just in case
                 print "Couldn't connect to server; aborting regression test"
+                # see if process is still alive
+                process.poll()
+                
+                if process.returncode is None:
+                    os.kill(pid(), signal.SIGKILL) # just in case
+                else:
+                    stdout_value, stderr_value = process.communicate()
+                    print "Stdout: %s" % (stdout_value)
+                    print "Stderr: %s" % (stderr_value)
                 sys.exit()
         else:
             client.transport.open()
